backend/requirements.txt
tensorflow
pydantic
pillow
mypy
asyncio
opencv-python
python-socketio